refactor: disclaimer helper 이름과 리뷰 메타데이터 정리
리뷰 P2 피드백에 맞춰 고지문 상수 파일명을 역할 중심으로 정리하고,
진행 이력과 리뷰 핸드오프 상태를 현재 결과에 맞게 갱신했다.

Constraint: 기존 진단 화면 동작과 고지 문구 자체는 바꾸지 않아야 함
Rejected: diagnosis.ts 이름 유지 | engine.ts와 책임 경계가 흐려짐
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: 리뷰 사이클 후속 수정은 REVIEW-N.md 상태와 Codex Response를 함께 갱신할 것
Tested: pnpm verify
Not-tested: 브라우저 수동 시각 검증
