+++ PROGRESS.md
@@ -17,6 +17,6 @@
 - [x] Vitest 회귀 테스트 2개 추가 (`ocr.test.ts`)
 
 ## 2026-04-06
-- [ ] 진단 결과 하단에 면책/기준일 고지문을 추가
+- [x] 진단 결과 하단에 면책/기준일 고지문을 추가
 - [ ] OCR 실패 시 빈 500 대신 원인별 오류 메시지를 반환하도록 보강
 - [ ] OCR 필드 매핑을 평가금액·매입가·현재가 기준으로 재정렬
