+++ REVIEW-1.md
@@ -1,0 +1,25 @@
+---
+cycle: 1
+branch: feat/8-disclaimer
+status: APPROVED
+p1_count: 0
+p2_count: 0
+---
+
+## P1 (must fix)
+없음.
+
+## P2 (optional — 반영 여부 Codex 판단)
+- [x] `src/lib/diagnosis.ts` → `src/lib/disclaimers.ts` 로 이름 변경. `engine.ts`가 진단 로직을 담고 있어 혼동 방지.
+  - import 경로, test 파일명, test 내 import 경로도 함께 변경
+- [x] `PROGRESS.md` 항목 `[ ]` → `[x]` (구현 완료 표시)
+
+## Codex Response
+- `src/lib/diagnosis.ts` / `src/lib/diagnosis.test.ts`를 각각 `src/lib/disclaimers.ts` / `src/lib/disclaimers.test.ts`로 변경하고 import 경로를 갱신했습니다.
+- `PROGRESS.md`의 2026-04-06 항목을 `[x]`로 갱신했습니다.
+- 변경 후 `pnpm verify`를 다시 실행했습니다.
+
+## Verdict: APPROVE
+
+p1 blocker가 없으므로 현 상태로 PR 올려도 됩니다.
+P2 반영을 원하면 수정 후 추가 커밋.
