+++ src/app/diagnosis/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation'
 import { ProblemCard } from '@/components/ProblemCard'
 import { AllocationBar } from '@/components/AllocationBar'
 import { ActionItem } from '@/components/ActionItem'
-import { DIAGNOSIS_DISCLAIMER_LINES } from '@/lib/diagnosis'
+import { DIAGNOSIS_DISCLAIMER_LINES } from '@/lib/disclaimers'
 import { SESSION_KEYS, DEFAULT_TARGET } from '@/lib/types'
 import type { DiagnosisResult } from '@/lib/types'
 import styles from './page.module.css'
